+++ src/main/resources/ha/shell/start_rsync.sh
@@ -17,8 +17,7 @@ function command_exists () {
 function install_rsync () {
 	
 	echo "$FORMAT_DATE : begin install rsync" >> $LOG_PATH
-	result=`whereis rsync | awk -F: '{print $2}'|sed "s/ //g"`
-	if [ -z "$result" ]; then
+	if [ ! -d "/usr/local/rsync"  ];then
 		echo "$FORMAT_DATE : install rsync......" >> $LOG_PATH
 		cd $HA_PATH/tools
 		tar -xzf rsync-3.1.2.tar.gz >> $LOG_PATH
@@ -28,9 +27,7 @@ function install_rsync () {
 	else
 		echo "$FORMAT_DATE : rsync has been installed. skip this step" >> $LOG_PATH
 	fi
-
-	result=`whereis rsync | awk -F: '{print $2}'|sed "s/ //g"`
-	if [ -z "$result" ]; then
+	if [ ! -d "/usr/local/rsync"  ];then
 		echo "rsync unable to install. Please see the installation log" >> $LOG_PATH
 		exit 1
 	else
@@ -62,17 +59,27 @@ function install_rsync () {
 
 function uninstall_rsync () {
 	echo "$FORMAT_DATE : begin uninstall rsync" >> $LOG_PATH
-	result=`whereis rsync | awk -F: '{print $2"x"}'|sed "s/ //g"`
-	if [ -z "$result" ]; then
+	if [ ! -d "/usr/local/rsync"  ];then
 		echo "$FORMAT_DATE : rsync is not installed. Skip this step" >> $LOG_PATH
 	else
-		$(`ps -ef | grep inotifywait |grep -v 'grep' | awk '{print $2}'  | xargs kill -9`)
-		killall rsync >> $LOG_PATH
-		cd $HA_PATH/tools/rsync-3.1.2
-		make uninstall >> $LOG_PATH
+		result_rsync=$(ps -e|grep rsync)
+		if [ -z "$result_rsync" ]; then
+			echo "$FORMAT_DATE : rsync is not Running. Skip this step" >> $LOG_PATH
+		else
+			killall rsync >> $LOG_PATH
+		fi
+		#cd $HA_PATH/tools/rsync-3.1.2
+		#make uninstall >> $LOG_PATH
 		if [ -e $HA_PATH/rsync_inotify/rsyncd.pid ]; then
 			rm -rf $HA_PATH/rsync_inotify/rsyncd.pid
 		fi
+		if [ -e $HA_PATH/tools/rsync-3.1.2 ]; then
+			rm -rf $HA_PATH/tools/rsync-3.1.2
+		fi
+		if [ -e /usr/local/rsync ]; then
+			rm -rf /usr/local/rsync
+		fi
+
 		echo "$FORMAT_DATE : uninstall rsync success" >> $LOG_PATH
 	fi
 }
@@ -80,7 +87,7 @@ function uninstall_rsync () {
 #安装inotify
 function install_inotify () {
 	echo "$FORMAT_DATE : begin install inotify" >> $LOG_PATH
-	result=`whereis inotifywait | awk -F: '{print $2}'|sed "s/ //g"`
+	result=`whereis inotify-tools | awk -F: '{print $2}'|sed "s/ //g"`
 	if [ -z "$result" ]; then
 		echo "$FORMAT_DATE : install inotify......" >> $LOG_PATH
 		cd $HA_PATH/tools/
@@ -92,10 +99,8 @@ function install_inotify () {
 		echo "$FORMAT_DATE : inotify has been installed. skip this step" >> $LOG_PATH
 		exit 0
 	fi
-	
 
-	
-	result=`whereis inotifywait | awk -F: '{print $2}'|sed "s/ //g"`
+	result=`whereis inotify-tools | awk -F: '{print $2}'|sed "s/ //g"`
 	if [ -z "$result" ]; then
 		echo "$FORMAT_DATE : inotify unable to install. Please see the installation log" >> $LOG_PATH
 		exit 1
@@ -107,13 +112,19 @@ function install_inotify () {
 
 function uninstall_inotify () {
 	echo "$FORMAT_DATE : begin uninstall inotify" >> $LOG_PATH
-	result=`whereis inotifywait | awk -F: '{print $2}'|sed "s/ //g"`
+	result=`whereis inotify-tools | awk -F: '{print $2}'|sed "s/ //g"`
 	if [ -z "$result" ]; then
 		echo "$FORMAT_DATE : inotify is not installed. Skip this step" >> $LOG_PATH
 	else
 		$(`ps -ef | grep inotifywait |grep -v 'grep' | awk '{print $2}'  | xargs kill -9`)
 		cd $HA_PATH/tools/inotify-tools-3.14
 		make uninstall >> $LOG_PATH
+		if [ -e /usr/local/inotify-tools ]; then
+			rm -rf /usr/local/inotify-tools
+		fi
+		if [ -e $HA_PATH/tools/inotify-tools-3.14 ]; then
+			rm -rf $HA_PATH/tools/inotify-tools-3.14
+		fi
 		echo "$FORMAT_DATE : uninstall inotifywait success" >> $LOG_PATH
 	fi
 }
@@ -123,14 +134,23 @@ function install(){
 	install_rsync
 	##安装inotify
 	install_inotify
+	##启动监听
+	$SHELL_PATH/switch.sh
+
 }
 
 function start(){
-	/usr/local/rsync/bin/rsync --daemon --config=$HA_PATH/rsync_inotify/config/rsyncd.conf
+	result=$(ps -e|grep 'rsync')
+	if [ -z "$result" ]; then
+		/usr/local/rsync/bin/rsync --daemon --config=$HA_PATH/rsync_inotify/config/rsyncd.conf
+	fi
 }
 
 function stop(){
-	killall rsync
+	result=$(ps -e|grep 'rsync')
+	if [ -n "$result" ]; then
+		killall rsync
+	fi
 }
 
 function uninstall(){
